1324 Add category refinements
