+++ app/pages/NavigatorDashboard/NavigatorDashboard.module.scss
@@ -76,6 +76,11 @@
   @media screen and (max-width: $break-tablet-p) {
     max-width: none;
   }
+  &.disabled {
+    background: $color-brand-light;
+    pointer-events: none;
+    cursor: default;
+  }
 }
 
 .searchFilterButton {
@@ -94,34 +99,6 @@
   }
 }
 
-.serviceFilters {
-  margin: 25px 0 60px;
-}
-
-.serviceTypeHeader {
-  color: $color-grey5;
-  font-weight: 600;
-  margin-bottom: 23px;
-}
-
-.serviceCheckboxList {
-  display: grid;
-  grid-template-columns: repeat(4, 1fr);
-  row-gap: 15px;
-  column-gap: 30px;
-  @media screen and (max-width: $break-tablet-l) {
-    grid-template-columns: repeat(2, 1fr);
-  }
-  @media screen and (max-width: $break-tablet-s) {
-    grid-template-columns: repeat(1, 1fr);
-  }
-}
-
-.checkboxItem {
-  display: grid;
-  grid-template-columns: auto 1fr;
-}
-
 .submitButtonRow {
   display: grid;
   justify-items: end;
@@ -129,6 +106,7 @@
 
 .submitButton {
   width: 200px;
+  margin-bottom: 12px;
   @media screen and (max-width: $break-tablet-p) {
     width: 125px;
   }
